requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
groq
python-telegram-bot
